# Backend/Swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateSchemaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/plan": {
      "get": {
        "operationId": "PlanController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Plan"
        ]
      },
      "post": {
        "operationId": "PlanController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Plan"
        ]
      }
    },
    "/plan/{id}": {
      "get": {
        "operationId": "PlanController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Plan"
        ]
      },
      "put": {
        "operationId": "PlanController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Plan"
        ]
      },
      "delete": {
        "operationId": "PlanController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Plan"
        ]
      }
    },
...
    "/service-tool": {
      "get": {
        "operationId": "ServiceToolController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ServiceTool"
        ]
      },
      "post": {
        "operationId": "ServiceToolController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceToolDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ServiceTool"
        ]
      }
    },
    "/service-tool/{id}": {
      "get": {
        "operationId": "ServiceToolController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service Tool ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ServiceTool"
        ]
      },
      "put": {
        "operationId": "ServiceToolController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service Tool ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceToolDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ServiceTool"
        ]
      },
      "delete": {
        "operationId": "ServiceToolController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service Tool ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ServiceTool"
        ]
      }
    },
    "/niche": {
      "get": {
        "operationId": "NicheController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Niche"
        ]
      },
      "post": {
        "operationId": "NicheController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NicheOptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Niche"
        ]
      }
    },
    "/niche/{id}": {
      "get": {
        "operationId": "NicheController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Niche ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Niche"
        ]
      },
      "patch": {
        "operationId": "NicheController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Niche ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NicheOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Niche"
        ]
      },
      "delete": {
        "operationId": "NicheController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Niche ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Niche"
        ]
      }
    },
...
  },
  "info": {
    "title": "ExBuySell - Api Documentation",
    "description": "",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserUpdateSchemaDTO": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "minLength": 4
          },
          "last_name": {
            "type": "string",
            "minLength": 4
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER",
              "MONITER"
            ]
          },
          "businessName": {
            "type": "string",
            "nullable": true
          },
          "contactName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "permissions": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "zip_code": {
            "type": "integer",
            "format": "int64"
          },
          "background": {
            "type": "string"
          },
          "is_online": {
            "default": false,
            "type": "boolean"
          },
          "password_hash": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string",
            "nullable": true
          },
          "otp_code": {
            "type": "string"
          },
          "approved": {
            "default": false,
            "type": "boolean"
          },
          "is_email_verified": {
            "default": false,
            "type": "boolean"
          },
          "is_phone_verified": {
            "default": false,
            "type": "boolean"
          }
        }
      },
      "CreatePlanDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 4
          },
          "description": {
            "type": "string",
            "minLength": 4
          },
          "duration": {
            "type": "string",
            "minLength": 4
          },
          "type": {
            "type": "string",
            "minLength": 4
          },
          "price": {
            "type": "string",
            "minLength": 4
          },
          "features": {
            "items": {
              "type": "array"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "required": [
          "title",
          "description",
          "duration",
          "type",
          "price"
        ]
      },
...
      "CreateServiceToolDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "image_path": {
            "type": "string",
            "minLength": 4,
            "nullable": true
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateServiceToolDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "image_path": {
            "type": "string",
            "minLength": 4,
            "nullable": true
          }
        }
      },
      "NicheOptionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}